fix tests

+++ src/StatefulT4Processor.BatchProcessor/Services/ProcessAndDeleteT4TemplatesService.cs
@@ -24,8 +24,11 @@ namespace StatefulT4Processor.TextTemplateZipProcessor.Services
 		{
 			foreach(var file in fileSystem.GetFiles(path))
 			{
-				t4TemplateHostWrapper.ProcessT4File(file, file.Replace(".tt", string.Empty));
-				fileSystem.DeleteFile(file);
+				if (file.EndsWith(".tt"))
+				{
+					t4TemplateHostWrapper.ProcessT4File(file, file.Replace(".tt", string.Empty));
+					fileSystem.DeleteFile(file);					
+				}
 			}
 
 			foreach (var directory in fileSystem.GetDirectories(path))
